+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 - Utils: Remove nagisa's Japanese word tokenizer and part-of-speech tagger
 - Utils: Remove ssg's Thai syllable tokenizer
 - Work Area: Remove Concordancer - Zapping Settings - Discard position information
+- Work Area: Remove Concordancer / Parallel Concordancer - Generation Settings - Sampling Method
 - Work Area: Remove Keyword Extractor - Generation Settings - Reference Files
 - Work Area: Remove N-gram Generator - Search Settings - Allow skipped tokens within search terms
 - Work Area: Remove Profiler - Generation Settings